[NETBEANS-3409] Missing popup menu for sql editor tabs
The Filesystems API compatibility module provides implementations for
Filesystem#getActions for several filesystems. Most filesystem
implementations are covered by the implementation in
AbstractFileSystemCompat, but this does not cover implementations
directly inheriting from Filesystem.

One such filesystem is BinaryFS.

In the SQL Editor the command window function (temp files for SQL
execution) is backed by a file in the SystemFileSystem, which delegates
in the end to BinaryFS:

SystemFileSystem->
  ModuleLayeredFileSystem ->
    BinaryFS

As BinaryFS does not implement getActions and no patch is provided for
this class, an AbstractMethodError is raised.

Instead of bailing out with an Error, the compatibility implementation
now returns an empty array to satisfy the Filesystem#getActions contract.