[NETBEANS-5031] - remove use of deprecated APIs in MiniEdit.java
I'm trying to remove the use of a deprecated API that hinders updating the Java version.

MiniEdit.java uses a really old deprecated API. This API prevents us from upgrade the Java
source version in the project properties file.

So step #1 is to remove the use of the old deprecated API.