[NETBEANS-5142] LSP Client creates excessive processes (#2589)
The caching of the LSP client that is rooted in project2MimeType2Server
does not work. In the WeakHashMap the keys are help by weak references
and the key is invalidated, once it is GCed. The URIs that are used
as keys become immediately eligible for GC and thus the cache is not
used at all.

To fix this, use a regular HashMap with WeakReference values.

The LSP Clients themselves are either strongly referenced from other
components or can be collected. To not create excessive start/stops of
servers, an additional strong reference is retained for a certain time
after the last bind request was issued. Only after the timeout has
occured, that extra reference is released and the bindings could become
eligible for GC.