Popup tab switcher should hide when another application steals focus
Fixes a problem I've run into several times on Mac OS - if another application steals focus
while the popup switcher is open, the switcher can wind up in a state where it is listening
twice, and is hidden as soon as you release the mouse button.

Fix is to add AWTEvent.WINDOW_EVENT_MASK to the set of AWTEvents listened for, check that
there really is no active window (to avoid confusion around heavyweight AWT popups) and
hide the popup in that case.