[NETBEANS-5304] CSS: Prevent UI blocking on undocumented properties
When the documentation for the target property could not be found, the
whole page of the CSS module was passed as help text. The rendering of
the HTML in combination with a popup menu caused problems on hide.

The documentation could be rendered, but on hide, the EDT is stuck in a
call to:

javax.swing.PopupFactory$MediumWeightPopup.hide()

Instead of rendering the whole documentation, the information "there is
no documenation" is rendered. While not optimal, it at least prevents a
heard freeze.