[NETBEANS-5351] NullPointerException from BreadcrumbsImpl
The caret for a text component can be null, to prevent a
NullPointerException, the existence of a caret needs to be checked
before the position is queried.

It is expected, that caret changes are only done on the EDT and thus
this code does not need to take multithreading into account.