Restore original format of m2:/ URLs modified by aa256463dcefb3b95bec6b7ac74e04a960015bc9.
m2:/ became a defacto API for the autoupdate mechanism. Both OpenJFX and
nb-javac rely on being able to be downloaded via the ".external" files
mechanism and third party modules could also rely on the exact syntax
of the URLs.

Instead of modifying the URL format, a new ad-hoc parser is introduced
to convert the m2 URL into maven coordinates.