Profiler lib native binaries: use ASF binaries with fix for NETBEANS-1428.
Use ASF-created binaries for Windows X86, Windows X64, MacOS X64, Linux X86 and Linux X64 which now includes the fix for NETBEANS-1428. For the rest of the platforms that were supported in the past (for example HP-UX, Solaris, Linux ARM) we keep these binaries in the distribution but they are no longer receiving bug fixes.