[NETBEANS-5698] Fix tabcontrol painting on MacOS retina displays (Aqua LAF)
This commit fixes scaling problems that occur in the window system tabs on the
Aqua LAF on Retina displays (2x HiDPI scaling). All MacBooks and other Apple
displays use retina scaling these days.

The fix is to use Graphics.fillRect instead of Graphics.drawLine; see the
comment in the code. This commit adds a new "drawLine" utility method in the tab
displayer UI classes for this purpose, and then does a plain search-and-replace
of "g.drawLine(" with "drawLine(g, ", with no other changes.

Note that on MacOS, standard UI panel borders are two device pixels thick on
Retina screens. So scaling everything up 2x on Retina displays is the right
thing to do, and looks consistent with native MacOS apps.

Before/after screenshots will be attached to the JIRA issue and GitHub Pull
Request.