Add the TreeView.setEntireRowClickable option.
If TreeView.setEntireRowClickable(true) is called, a left click, right click, or drag operation in the empty area on the right-hand side of a node will work as if done on the node itself. Note that for left clicks, this is already the default behavior on FlatLAF.

Since some TreeView usages in the IDE may have important context menus that show up if you right-click the empty area, the option is left off by default, and is currently unused. But it is useful in platform application.