Make Glassfish 6.0 use the same JDKs as before with max 1.8. Set possible JDKs for GF 6.1 to 1.8, 11 and 16.
I'm not sure if GF 6.1 works on 1.8 as it's not clear from the
documentation, but included nevertheless. 
Officially only JDK 11 is supported on GF 6.1, but 16 mainly works so
JDK 16 is there too.

Removed all other intermediate JDKs.

Controlled that GF 6.1 correctly launches from Netbeans running
on JDK 16 with Glassfish platform set to JDK 16.