[Netbeans-5734] Fix extraction of jvm-options from domain.xml and revert add-opens directives
The additions of the "add-opens" arguments for the JVM are not
necessary if the jvm-options are correctly extracted from the
domain.xml.

It was observed, that glassfish cleanly starts with JDK 11, so something
strange was going on in the way NetBeans started it. The issue can be
tracked down into ServerTasks#appendOptions. In that method variable
substitution and proxy adjustments are done. The arguments are stored
in a combination of Map argumentName -> value and and a List for the
order of the arguments. The issue is the Map, as this limits each
argumentName to be present at most once, but there are multiple
add-opens directives. Only the last definition survived.