Fixes wrong detection message when seelcting a GF6.2 location.
When adding a new server, after selecting "Glassfish Server" and
pressing next, when selecting an install location, the GUI searches for
a valid instance by comparing all possible instances (versions) with 
the one found in the location specified.

..modules.glassfish.common.wizardsServerWizardIterator::isValidInstall
..modules.glassfish.common.ServerDetails::isInstalledInDirectory

If the Glassfish version is 6.2, it shows it has found 6.1, but then 
goes on with the 6 version, as the 6.1 instance says it's 6.

org.netbeans.modules.glassfish.common.ServerDetails::getVersion

Now 6.1 says it's 610 and 6.0 says it's 600 as other versions do.