[NETBEANS-2360] Auto-detect HiDPI and anti-aliasing settings on Linux.
This commit modifies the NetBeans IDE launcher scripts to automatically handle two display-related conditions on Linux:
(1) Automatically enable the -Dawt.useSystemAAFontSettings system parameter if necessary to enable anti-aliased text. Veteran NetBeans users have known about this setting for a long time, but after a lot of research, I believe I have found the specific condition in which it is necessary, namely on the K desktop environment.
(2) Automatically detect cases where the user has enabled 2x HiDPI scaling, and set the GDK_SCALE environment variable accordingly so that OpenJDK will handle this accordingly. This seems to work on both KDE and Gnome. HiDPI scaling works best on FlatLAF (which is now the default LAF for NetBeans on all platforms), but seemed to work tolerably on the default GTK LAF as well. See the in-code comments for more details. (There are two copies of the new code: one in the IDE launcher script, and one in the launcher script for NetBeans Platform applications.)