Prevent background tasks working on LSPBindings to be scheduled multiple times
Observation: Removing 100 characters from a typescript file, would lead
to 100 invocations of the background tasks. If the removal is done fast,
there will be multiple invocations for the same task, which are first
delayed for 500ms and then are all executed one after another.

This is unnessary, as a queued task will get the latest update anyway.

The idea to fix this is to save which tasks are already enqueued and if
they are already enqueued, skip the additional invokation.