Better implementation of PSR-4 Namespace rules
Consider src to be the base directory and the non-matching subnamespaces after src is encountered as the namespace prefix.
Implementation is rought, for it to be correct, i would need to add a configuration in the project to set the namespace prefix, fetch configured namespace prefix from the project in this Checker and remove the appropriate leading part of the namespace. I don't know how to do that.