[NETBEANS-5931 part 2/2] Use configureDefaultRenderingHints throughout the codebase.
In this commit, I went through all occurrences of the following strings in the
NetBeans codebase, and used GraphicsUtils.configureDefaultRenderingHints in
place of previously existing, duplicated logic:

  awt.font.desktophints
  shouldAntialias
  gtkShouldAntialias
  swing.aatext

This should unify antialiasing behavior throughout the IDE. The new logic may
not be exactly the same in all cases, since there are many small variations in
how different parts of the IDE applies rendering hints. But arguably the
rendering hints should be the same everywhere.

The text editor infrastructure has its own hints-related code, which is not
touched here.