VaniallaPartialReparser reports incorrectly reparsed files
The verification step of the VanillaPartialReparser serialises the
reparsed tree to a string and comparse these. The issue is, that the
toString result of a CapturedType contains the sequence
"capture#NUMBER". Number is the hashCode of the mirror. As hashCode
is not overwriten, this is more or less a random number and thus
meaningless for the tree comparisson leading to invalid incorrect
partial reparsing reports

This normalises it to a plain capture.