CheckRegexp hint causes all lines with regexp shown with a warning
The severity of the CheckRegexp of VERIFIER causes all lines in source
files containing a regexp to be marked as warning. This is invalid, as
the regexp are perfectly ok.

The severity is reduced to HINT. That way the hint is readily available
when the regexp is under the cursor, but does not clutter the whole
file.