[NETBEANS-5734] Add details for Glassfish 6.2.1 server and improve version detection
With Glassfish 6.2.1 support for JDK 17 was added, so meta data for
that version was added.

The version detection was improved in the following way:

If an exact match is found, that entry is used. Else the list of
server definitions is iterated from highest to lowest. The highest
definition, where the version number is lower than or equal to the
target version is used to configure the server.

As an example: If you try to configure Glassfish 6.2.2, it will not
find an exact match. Instead 6.2.1 is matched. Prior to this change,
the configuration for 6.0 would have been chosen.