[NETBEANS-6115] UTF-8 input is broken in Maven project.
Chars were written as bytes to a stream, ignoring the encoding.
By wrapping the stream in a writer, chars can now be encoded properly.

Chose to not specify the encoding so that the default can be used or
overwritten by -Dfile.encoding=UTF-8.

JDK 18+ uses UTF-8 as default, it will be probably best practice
to set the encoding to UTF-8 for most JDK 17 projects too, while using
the new 'native.encoding' property for edge cases.