Update gradle to 7.3 and rewrite netbeans-gradle-tooling in java
The gradle binaries are only distributed from their own download site.
From the "gradle-bin" distribution the wrapper can be extracted to be
then distributed with netbeans.

The netbeans-gradle-tooling that is executed as part of the build war
rewritten in java, as the updated groovy version bundled with gradle 7.X
creates binaries, that are not compatible with previous versions anymore
and building with gradle 6.X fails on JDK 17. Using Java makes it easier
to build compatible binaries.