Upgrade to Apache Maven 3.8.4: Adjust to DefaultModelBuilder changes
In maven 3.6 the method

ModelBuilder#build(ModelBuildingRequest, ModelBuildingResult)

at the end of the call to

ModelBuilder#build(ModelBuildingRequest)

In 3.8.X the call chain was modified, so a private method of
DefaultModelBuilder is called instead of the two parameter public
method.

The implementation in NBModelBuilder called first the super
implementation. So in essence the sequence in the past was:

- NBModelBuilder#build(ModelBuildingRequest)
  - DefaultModelBuilder#build(ModelBuildingRequest)
    - NBModelBuilder#build(ModelBuildingRequest, ModelBuildingResult)
      - DefaultModelBuilder#build(ModelBuildingRequest, ModelBuildingResult)
      - [NB Extensions]
  - [NB Extensions]

The modified implementation ensures, that the post processing happens
independend from the call chain.