Stabilize ShellConstructorTest and  RunJarStartupArgsTest on JDK 9+
---------- RunJarStartupArgsTest ----------

Without maven.compiler.source and maven.compiler.target set, they are
defaulted to 1.5. This will fail on recent JDKs. Instead use the java
version from the running JDK, which can be assumed to be a save default.

---------- ShellConstructorTest ---------

The test is flaky and it was found, that sometimes the maven CLI version
is reported to be NULL. Closed analysis showed, that this is caused by a
IOException, referencing to an unexpected end of zlib input stream.

The issue seems to be, that the ArchiveURLMapper caches jars and the
test in rapid succession deletes jars and recreates them.

This change moves dummy JAR creation from per Test to per Testcase.
This shoud stabilize the test.