Follow LSP specification for null value for startCharacter in FoldingRange
The LSP specification specifies for the startCharacter:

The zero-based character offset from where the folded range starts. If
not defined, defaults to the length of the start line.


In contrast to the specification the first character was choosen as the
start of the fold, hiding for example the function name.