Fix LogContext - logStartIndexer and logFinishIndexer need to be called as pairs
The logStartIndexer is called once per mimeType, while the
logFinishIndexer calls are done once per source. As logFinishIndexer
pops the current indexer of the indexer stack, the checks, that verify
balancing fail and raise warnings.

Instead of calling logFinishIndexer from UserTask, it needs to be called
from the invoking method.