[#3593] Don't assume contract StructureItem#getElementHandle is never null is held
There are several places in the NetBeans codebase, that violate the
contract of StructureItem#getElementHandle and return null. This code
fails in that case.

An assert is added, so that when testing further places, that fail the
assert can be identified, but it does not hit users.

In addition the FileObject are compared via #equals and not by identity.