[#3728] Only enable VerifyPartialReparse if assertions are enabled
It was observed, that NetBeans can create an excessive number of source
dump files in situations, where the partial reparse verifier bails out.

The VanillaPartialReparser already suppresses dumping of source files
in case of exceptions if assertions are not enabled. The same logic is
now applied to the activation of VerifyPartialReparse.
VerifyPartialReparse has no side effects and thus can be disabled if
dumping should be turned off.