Fix naming of application for GnomeKeyring and LibSecret
The application names for the Gnome Keyring and LibSecret integrations
are derived from the name of the application. That name contains the
release version, so the passwords are lost on each update.

Instead of deriving the name from the application name, GnomeKeyring
and LibSecret are brought in line with the existing implementation of
KWallet and mac OS, where the name is hardcoded.

As the names for GnomeKeyring and LibSecret have to be change at least
once more, they are unified to the KWallet name (which might be
accessed by LibSecret)

Closes: #3702