SQL History loss: Reimplement atomatic saving by writing to tempfile and moving to target (#3868)
It was observed, that sometimes the sql history was lost and it is
assumed, that writing the new history began (emptying the file as the
first step) and was not completed. By using a temporary file, the file
is either written completely or not. It is assumed, that a move
operation inside the same directory is atomic on all OSes.