Enable parsing of object structures, where the prototype contains complex properties
Complex properties in this case means properties, that are defined not
only by a variable, but a property of another object or a function call.

The case:

object.prototype[variable] = function() {}

was already covered, but the cases:

object.prototype[object2.property] = function() {}
object.prototype[dummyFunction()] = function() {}

caused a the whole object structure to be invalid.