LSP Client: Log unhandled requests and notifications from LSP server and return empty result
This needs a real implementation where LSP Clients handle these kinds
of requests, but at least the typescript language server uses a request
to indicate a rename (_typescript.rename), but is not interested in the
result of the operation.

It is expected that normally lsp server first ensure the client can
handle the request and don't unconditionally send potentially unhandled
requests.