Solaris SPARC is a dead platform.
No more Java support - JEP 381

In 2017, Sun shutdown the SPARC dev. teams and stopped future work on the processor.

There will be additional PRs to	cleanup	the remaining SPARC bits. But, this removes platform support which is
a fundamental enabler.

It's makes me sad to have to do this but it's the right thing to do.