Fix maven groupId completion
MavenIndexer dropped the caching of the allGroups and rootGroups
properties in the index:

https://github.com/apache/maven-indexer/commit/c231db602b5cea9ed20eba617de5a5a846ef5faa

After the index is loaded there are two options:

a) rebuild the groups list from the index
b) cache the group list after indexing

NetBeans expects the index to be readily be available and thus option
a is not an option (rebuilding the group list runs over the full index
and is in the minute range). As there are only a few entrypoints that
modify the index caching the group list looks like a viable option.

Closes: #4102