Invoke indexer in RepositoryUpdater grouped by mimepath of file
While analysing this it was found, that the GsfParserFactory#createParser
returned null as the parser for snapshots. Looking deeper, showed, that
the #createParser method was invoked with snapshots of differing mimetypes.

This works, but is a violation of the contract in ParserFactory#createParser:

> It is guaranteed that all snapshots in the collection will be
> of the same mime type and it will be the mime type, which this factory
> was registered for (ie. in MimeLookup)

If the snapshots are inconsistent, parsing still works, but its performance
suffers. To fix this the call to ParserManager#parse in RepositoryUpdater
is split. The scanned sources are grouped by their mimepath and each group
is individually passed.

The observed improvements are:


Indexing Variant 1 (tour-of-heroes)

Before: ~125s
After:  ~105s

Indexing Variant 2 (tour-of-heros with Indexability Filter)

Before: ~52s
After:  ~38s

The Indexability Filter blocks scanning for all paths with paths containing
node_modules and where the indexer is in the list: js, angular, requirejs,
knockoutjs, TLIndexer, tests.