Limit ValidateClassFilesTest to scan only JARs referenced from modules (#4775)
With the addition of jakarta ee 10 API jars it was found, that there is
an implicit assumption, that all JARs not referenced by modules have
to be JDK 8 compatible. In the case of jakarta ee 10 this is an
invalid assumption as it explicitly does not support JDK 8. The module
that introduces the binaries correctly declares a JDK 11 dependeny, but
at runtime the module <-> binary association is lost.

To prevent false positives, the test must be limited to JARs, that are
directly referenced from modules.