Remove support for Windows 95/98/XP
There is very minimal support for Windows 95/98/XP in the source tree. Nothing that couldn't
be handled by the generic Windows code. Also, these platforms are so old, it would be very difficult
for the Netbens teams to fix a problem, if it was even reported.

In addition, Windows 95/98/XP has long been deprecated. Look here for references:

https://www.java.com/en/download/help/winxp.html

https://www.oracle.com/java/technologies/javase/products-doc-jdk8-jre8-certconfig.html

So this code change cleans up (i.e. removes) what little specific Windows 95/98/XP support we have.

It would be helpful to clean up other deprecated platforms.