fixes an ambiguous method reference.
The compiler can't decide if this is an ExceptionAction or a Runnable
and marks the line as an error.

We can help the compiler by casting to Runnable or by using a lambda
expression. Cast is more compact.

(Without this change the build would fail as soon --release is set,
the editor did mark it as error even before that)