[NETBEANS-1914] Fix startup problems on Windows after moving NetBeans install directory.
On Windows, renaming or moving the NetBeans installation directory caused errors on startup, due to references to old paths in all-layers.dat in the cache directory. The bug did not seem to happen on Linux or MacOS; I'm not sure why. See https://issues.apache.org/jira/browse/NETBEANS-1914 for an example stack trace.

The solution in this patch is to include the netbeans.home path in the all-checksum.txt file, on Windows. This will invalidate the cache if the installation directory is changed. Some tests had to be modified to permit this change.