move webcommon job to JDK 11 and fix tests where needed.
JsDocCodeCompletionTest can't use getTestFile() in overwritten method
since the method fails the test if the file does not exist. The
Test file logic in CslTestBase::assertDescriptionMatches has to be able
to call exists() on a potentially not existing file when it is trying
to find JDK version specific (golden) test files.