tomcat5: Enable usage of javac release and remove internal class use (fix JDK8 and JDK17+)
The tomcat5 module used the internal classes:

com.sun.org.apache.xml.internal.serialize.OutputFormat
com.sun.org.apache.xml.internal.serialize.XMLSerializer

to modify the tomcat configuration. For this to work the usage of the
javac release option had to be disabled as only exported classes are
then accessible. This is done by specifying bootclasspath.prepend.

Having this masked a different issue which rendered the tomcat module
JDK 9+. The method:

URLEncoder#encode(String s, Charset enc)

was used, which was introduced with JDK9. The alternative

URLEncoder#encode(String s, String enc)

can be used as a drop-in replacement.

Closes: #5134
Closes: #5154