Upgrade bundled maven to 3.8.7.
Maven dist dropped commons-io as dependency.

Fix building of ueber-jar for netbeans-eventspy:

The shading applied by jarjar tool fails on JDK 8+. In consequence the
classes from simple json are renamed, but the references to these
classes are not fixed. This works as long as a copy of simple json is
also on the classpath.

Co-authored-by: Matthias Bläsing <mblaesing@doppel-helix.eu>