[NETBEANS-4123] Implemented handling large strings for Java 9+
In the debugger (+7 squashed commits)
Squashed commits:
[5e450181470f] [NETBEANS-4123] Lifted exceptions from ShortenedStrings.isLittleEndian

Waiting for this change to be reviewed before squishing
[fc97d860fd01] [NETBEANS-4123] Checkpoint before extracting exceptions...
[e6cedcda8c5d] [NETBEANS-4123] Moved sync blocks around as per @matthiasblaesing
[1b1da080614d] [NETBEANS-4123] Now caching whether or not a VM is little endian

Check is now deferred to when it is needed instead of right away.
[4f52c498989c] [NETBEANS-4123] Finished implementing the handling of long strings

Renamed 'compressed' things to 'compact'
Tested with a long UTF16 string on little endian. It detects the byte order properly
[e8d001f2fd5e] [NETBEANS-4123] Apply suggestions from code review

Will apply other changes and testing in a bit since you can't do those in GitHub

Co-authored-by: Matthias Bläsing <mblaesing@doppel-helix.eu>
[2caef567d9a3] [NETBEANS-4123] Initial implementation of handling large strings

in Java 9+ where the String class uses a byte[] instead of a char[]