Support recursive CSS property definitions
The grammar for calc property values failed to parse embedded var and
calc calls correctly. The issue is the work-around introduces as
function-content, which basicly prevents nesting any expression that
contains braces.

This changes the grammar parsing result from a tree to a directed graph,
which can contain cycles. This needs to taken into account by useers of
the parsed grammar.

An example of such a cycle is:

calc-fn -> calc-sum -> calc-product -> calc-value -> calc-sum
              ^                                        |
              +----------------------------------------+

Closes: #5025