Remove usage of Thread#suspend and Thread#resume from INSANE
The methods Thread#suspend and Thread#resume are deprecated. The only
call-sites in INSANE in ScannerUtils are private methods, that are only
called from commented out code. The commented out code is in that state
sine 2007 and no longer history is available.

Instead of trying to fix this, the methods and the commented code are
removed. The documentation of the surrounding method is adjusted to
not promise things, that are not true for more than a decade.