Ensure indention is handled in a predictable order
It was observed that the test

org.netbeans.modules.php.editor.js.JsFormatterEmbeddedTest.testEmbeddedMultipleSections1()

was flaky. It failed randomly and investigation showed, that when it
happend the invocation order of the indenters/formatters differed:

Fail: 
    [junit] HtmlIndentTask#reindent
    [junit] HtmlIndenter#reindent
    [junit] CssIndenter#reindent
    [junit] JsFormatter#reformat
    [junit] PHPFormatter#reformat

Ok:
    [junit] HtmlIndentTask#reindent
    [junit] HtmlIndenter#reindent
    [junit] JsFormatter#reformat
    [junit] CssIndenter#reindent
    [junit] PHPFormatter#reformat

The order indeed can change as the items to be processed are taken from
a HashSet (thus might be affected by memory location), put into an
ArrayList and then sorted. The sorting is only based on the length of
the language path and for two paths with identical length no total
order is defined. This is fixed by introducing a second sort factor:
If the length of the path is identical, the paths are alphabetically
sorted, based on the mimetype.

This fix exposed another problem in the indenter implementations. There
are already reordering steps implemented (PHP, ruby, ...), but CSS and
JS also need to be special cased.