bump php job to test on JDK 11 and fix failing tests
fix php folding test on JDK 11+ by opening the document in advance.

FoldingScanner#folds calls source.getDocument(false) and receives non
null values on JDK 8. On JDK 11+ however the document is null, which
leads to test failures in FoldingTest#testPHPTags.

This is likely a race condition which is more likely to occur on
JDK 11+ since the test can be brute forced to passing on linux if
restarted often enough.