Maven indexing: Retain download index setting even if index setting is disabled
When the maven general indexing is disabled, the setting for the index
download is lost and forced to be false. Instead of loosing the value
a new preferences getter is introduced, that calculates the effective
value based on the two components:

- general indexing
- download index

Only if both settings are true, the effective value will be true.