try to infer compiler plugin version from active maven version.
The plugin version query returns the version of the embedded maven
distribution and not the verison of the active distribution.

This worked fine until maven updated the versions of their implicitly
provided lifecycle plugins (which is a good thing), then it broke.

We can workaround this problem by comparing maven versions, since
we can infer the plugin version from it under certain conditions.

Fixes the release option and the module-info hint.