html.validator: Remove dependency on log4j
This commit modifies the patched version of the html.validator not to
rely on log4j anymore.

There is no security gain estimated. This commit exists only to silence
"security" teams, that have some strange interpretation of security.
They assume, that you can deduce from the presence of a dependency/
library, whether or not there is a security problem. There is no
evidence, that any of the CVEs against log4j 1 apply here.