don't block in NBM wizard + skip snapshots + more.
the new NBM project wizard could potentially block after finish is
pressed, waiting for the indexer to provide complete results for
the archetype version query.
This is unnecessary since archetypes rarely change and the default
is usually the correct version even if the index isn't ready yet,
so we are ok with partial results.

Further, it didn't filter out snapshot releases, which could be
potentially in the local repository too.

It could even downgrade the archetype version by accident under some
circumstances when indexing got previously canceled.

+ fixed some deprecations and NPEs in maven hints

+ bumped archetype and plugin version defaults to latest