Stabelize html lexer in broken cases
Faced with a structure like this:

"<Jm/  a=q</>>V  />G>"

The html lexer hit an assertion as it tried to lex "a=q" as an
attribute value. The sequence "<Jm/" at first looks like the start of
a tag, but that is discarded once the "/" is read. Yet the lexer state
is not reset to INIT, but ISP_TAG_X.

Attribute lexing assumes, that a tag was lexed first and that that
information is still present, which not the case here.