basic auto completion for maven version properties.
It is somewhat common to use properties for artifact versions, esp
when several artifacts share the same version.

This implements basic auto completion for properties which are used
as version field.

In contrast to hints, completion is implemented in the grammar
and doesn't have the full context. This will only take properties
into account which are in the same pom file as their usage.