[JS] Improve resolving this in ClassElements and in arrow functions
Inside initialisation statements:

class Test {
	var1 = 123;
	var2 = this.var1;
}

this needs to evaluate to the surrounding class, to handle this
correctly, search for the right definition of "this" needs to start
at the surrounding object, not the declaration scope of the wrapping
function.

In arrow functions this is not redefined, but inherited from the outer
scope. In this sample:

class Foo {
	bar = {};
	
	test(someArray) {
		this.bar = {}; // Mark
		someArray.forEach((val) => {
			this.bar[val] = 1;  // Mark
		});
	}
}

this in the lines with the "Mark" comment is an instance of `Foo` and
thus it is always the same bar. The this scanning needs to skip arrow
functions.

Closes: #5740
Closes: #4376