Cleanup the use of Double primitive wrapper.
Cleanup the use of Double primitive wrapper. As of Java 9, the VM can do a better job of handling the conversion from a Double primitive to a Double Object or vice versa.

This is an extension of the work done in #2498.

No tests have been touched as those expressly need to manage the conversion.

A couple of places I also cleaned up the Float.