On JDKs without JavaFX/OpenFX libs.javafx has unsatisfiable package dependency
With this commit:

commit 15811ec850cc3cfe50eeaffd96c2eae1065ccb9c
Author: Svata Dedic <svatopluk.dedic@oracle.com>
Date:   Tue Apr 18 14:10:14 2023 +0200

    Allow java and (system) package dependencies for eager modules.
    Provider autoloads with failed dependencies have problems reported.

dependency checking was improved and in turn module configurations that
worked by luck failed to load.

For libs.javafx the situation is trivial on JDKs that bundle
JavaFX/OpenJFX. In these cases the JDK provides the OpenJFX classes,
which is detected by the module system and the token
org.openide.modules.jre.JavaFX is automatically provded. On other JDKs
OpenJFX is loaded from a set of org.netbeans.libs.javafx.<PLATFORM>
modules onto the classpath and these modules provide the mentioned
token.

Before this change libs.javafx had this manifest entry:

OpenIDE-Module-Package-Dependencies: javafx.application[Application]

With the improved checking this requirement fails on JDK not bundling
JavaFX/OpenJFX. For these cases both the fragement host
org.netbeans.libs.javafx and one of the provider modules, for example
org.netbeans.libs.javafx.linux would need to be loaded together to
be able to satisfy the dependency.

This PR removed the package dependency as it provides no additonal
value, as the presence of the token org.openide.modules.jre.JavaFX
reports presense of native support.

Closes: #6004